New SwiftUI `DemoHostingController` parent class
